There was a big problem related to using getIt
Refactor BlocProvider initialization in AppRouter

Refactor the initialization of `PersonCubit` and `GenreCubit` in the `AppRouter` class to ensure that the `fetchTrendingPersons` and `fetchGenres` methods are called after the cubits are created. This change addresses potential issues with using `getIt` directly in the `create` method. Added TODO comments for further investigation and