Initialize and fetch data in cubit constructors
Refactor the `GenreCubit` and `PersonCubit` constructors to automatically fetch genres and trending persons respectively when the cubits are initialized. This change removes the need to manually call `fetchGenres` and `fetchTrendingPersons` in the `AppRouter` and ensures that the data is fetched as soon as the cubits are created. Additionally, commented out the old implementation in `AppRouter` and added a todo note for further consideration.